requirements_snapshot.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiolimiter==1.2.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.64.0
anyio==3.7.1
APScheduler==3.11.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
boto3==1.34.143
botocore==1.34.162
cachelib==0.13.0
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
contourpy==1.3.0
cryptography==45.0.6
cycler==0.12.1
dash==3.2.0
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
docstring_parser==0.17.0
docx2txt==0.9
ecdsa==0.19.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.2.0
faiss-cpu==1.12.0
fastapi==0.104.1
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.59.1
frozenlist==1.7.0
fsspec==2024.6.1
genai==2.1.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.111.0
google-cloud-bigquery==3.36.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.30.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipython==8.18.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
kombu==5.5.4
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.74
langchain-huggingface==0.3.1
langchain-openai==0.3.30
langchain-text-splitters==0.3.9
langdetect==1.0.9
langsmith==0.4.14
limits==4.2
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgspec==0.19.0
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==2.1.2
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
openai==0.27.10
openpyxl==3.1.2
ordered-set==4.1.0
orjson==3.11.2
packaging==24.2
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
ProjectAPIs @ file:///home/ec2-user/apps/JobShuffler/QualityFlowPythonSDK%20%282%29.zip#sha256=191f857cacc56390022112b4277a035c783ad3bcd1b2975b0c717ba97a7f421a
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.10.1
pydantic_core==2.14.1
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.3
pypdf==6.0.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-http-client==3.3.7
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.0
requests-file==2.1.0
requests-toolbelt==1.0.0
retrying==1.4.2
rich==13.9.4
rpds-py==0.27.0
rsa==4.9.1
s3transfer==0.10.4
safetensors==0.6.2
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.13.1
sendgrid==6.10.0
sentence-transformers==5.1.0
shapely==2.0.7
simple-salesforce==1.12.4
six==1.17.0
slack_bolt==1.23.0
slack_sdk==3.36.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
starkbank-ecdsa==2.2.0
starlette==0.27.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.3.3
tokenizers==0.21.4
torch==2.4.0+cpu
tqdm==4.66.4
traitlets==5.14.3
transformers==4.55.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.20
uvicorn==0.24.0
valkey==6.1.1
vine==5.1.0
waitress==3.0.2
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
WTForms==3.2.1
xlsxwriter==3.2.5
yarl==1.20.1
zeep==4.3.1
zipp==3.23.0
zstandard==0.23.0
